cloud-functions/ack-endpoint/requirements.txt
functions-framework==3.*
google-cloud-logging>=3.4.0
google-cloud-firestore>=2.7.3
